Adapt code to love2d 11.0
This introduces a new colour component range mode, which defaults to `"native"`. The other possible value is `255`. In native mode, the colour components used for styles use the same range as the version of LÖVE being used, be it 0-255 or 0-1. In 255 mode, the colour components use always the range 0-255 regardless of the version running.